+++ test/reindex_test.rb
@@ -39,6 +39,18 @@ class ReindexTest < Minitest::Test
     assert_search "product", ["Product A"]
   end
 
+  def test_async_non_integer_pk
+    skip if !defined?(ActiveJob)
+
+    Sku.create(id: SecureRandom.hex, name: "Test")
+    reindex = Sku.reindex(async: true)
+    assert_search "sku", [], conversions: false
+
+    index = Searchkick::Index.new(reindex[:index_name])
+    index.refresh
+    assert_equal 1, index.total_docs
+  end
+
   def test_refresh_interval
     reindex = Product.reindex(refresh_interval: "30s", async: true, import: false)
     index = Searchkick::Index.new(reindex[:index_name])